% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{author_fullname}
\alias{author_fullname}
\title{Compute author full name from given & name fields}
\usage{
author_fullname(given, name)
}
\arguments{
\item{given}{given name field}

\item{name}{name field}
}
\description{
Compute author full name from given & name fields
}
