% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{cache_file}
\alias{cache_file}
\title{Cache loader from file
Expose load() & save() function to simply load from cache}
\usage{
cache_file(name, ttl)
}
\arguments{
\item{name}{file name (without path)}

\item{ttl}{Time to live in seconds, if NULL default from config is used}
}
\description{
Cache loader from file
Expose load() & save() function to simply load from cache
}
