% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{format_citation}
\alias{format_citation}
\title{Format publication to a textual citation}
\usage{
format_citation(p, authorlist = "all")
}
\arguments{
\item{p}{data.frame row of a publication table}

\item{authorlist}{all, first or "first_last"}
}
\value{
character string of citation
}
\description{
Format publication to a textual citation
}
