% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{install_teams}
\alias{install_teams}
\title{Install teams list loaded from server in config}
\usage{
install_teams(teams)
}
\arguments{
\item{teams}{list}
}
\description{
Install teams list loaded from server in config
}
