% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_records}
\alias{load_records}
\title{Load all records using pagination}
\usage{
load_records(page_size)
}
\arguments{
\item{page_size}{integer, number of records by page}
}
\description{
Load all records using pagination
}
