% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_records_page}
\alias{load_records_page}
\title{Load Pubweb Records using pagination}
\usage{
load_records_page(offset, page_size, with_team = FALSE)
}
\arguments{
\item{page_size}{integer number of records by page}

\item{with_team}{logical if TRUE add teams list in results}

\item{page}{integer page number to load}
}
\description{
Load Pubweb Records using pagination
}
