% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_with_cache}
\alias{load_with_cache}
\title{Load RDS with cache}
\usage{
load_with_cache(file, ttl)
}
\arguments{
\item{file}{file name to load}

\item{ttl}{int time to live in seconds}
}
\description{
Load RDS with cache
}
