% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{pub_to_bib}
\alias{pub_to_bib}
\title{Create bibentry from a publication row}
\usage{
pub_to_bib(p)
}
\arguments{
\item{p}{row of a publications dataframe}
}
\value{
bibentry
}
\description{
Create bibentry from a publication row
}
