% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{pubweb_add_team_columns}
\alias{pubweb_add_team_columns}
\title{Extract teams attached of each publication and add a corresponding logical column for each team}
\usage{
pubweb_add_team_columns(.db, recode = TRUE, col_prefix = "team_")
}
\arguments{
\item{.db}{data.frame() of publication with at least columns \code{id} and \code{equipes}}

\item{recode}{if TRUE team label will be used (from \code{recode_team}) instead of team id}

\item{col_prefix}{prefix to add to the team label (e.g. "team_1", "team_2")}
}
\value{
given data frame with one extra column by team
}
\description{
Extract teams attached of each publication and add a corresponding logical column for each team
}
