% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{pubweb_authorlist}
\alias{pubweb_authorlist}
\title{Create list of authors (coma separated list of authors)}
\usage{
pubweb_authorlist(authors)
}
\arguments{
\item{authors}{authors column}
}
\description{
Create list of authors (coma separated list of authors)
}
