% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{pubweb_authors}
\alias{pubweb_authors}
\title{Extract authors list from th database}
\usage{
pubweb_authors(.db, with = c())
}
\arguments{
\item{.db}{database loaded from \code{\link{pubweb_load_db}()}}

\item{with}{columns to keep}
}
\value{
data.frame()
}
\description{
Extract authors list from th database
}
\details{
\describe{
\item{id}{Publication id}
\item{order}{Order of the authors in the publication}
\item{name}{First Name of the author. Depending of the data source, this field can contain the full name or the family name}
\item{given}{Given name of the author, sometimes not provided}
\item{iplesp}{This author is recognized as an Iplesp author with corresponding affiliation}
\item{wild}{This author is part of Iplesp, but affiliation is not mentionning ig}
\item{affiliation}{Affiliation provided, several affiliations has been concatenated with \code{|} character}
\item{hal}{halId of the author, rarely provided}
\item{orcid}{Orcid number of the authors, as provided by the metadata source (pubmed, wos,...)}
\item{corresp}{Authors is corresponding author}
\item{group}{Author is a group}
\item{equal}{Equal authorship with the adjacent author (first or last)}
}
}
