% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pubweb_cache}
\alias{pubweb_cache}
\title{Get Path of cache file for a given name}
\usage{
pubweb_cache(name)
}
\arguments{
\item{name}{character vector}
}
\description{
Get Path of cache file for a given name
}
