% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{pubweb_extract_hal}
\alias{pubweb_extract_hal}
\title{Create data.frame with correspondance with HAL ids and iplesp id}
\usage{
pubweb_extract_hal(.db, with_aliases = TRUE)
}
\arguments{
\item{.db}{database}

\item{with_aliases}{if TRUE include HAL aliases in the list}
}
\description{
Create data.frame with correspondance with HAL ids and iplesp id
}
