% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pubweb_load_from_server}
\alias{pubweb_load_from_server}
\title{Load Pubweb databases froms server}
\usage{
pubweb_load_from_server(ttl = NA)
}
\arguments{
\item{ttl}{int cache time (if 0 will disable cache)}
}
\description{
Load Pubweb databases froms server
}
