% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pubweb_load_teams}
\alias{pubweb_load_teams}
\title{Load team list from server (cached)}
\usage{
pubweb_load_teams(ttl = NA)
}
\arguments{
\item{ttl}{Time to live, if NA default ttl in config will be used, O will disable cache}
}
\description{
Load team list from server (cached)
}
