% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hceres.R
\name{pubweb_recode_hceres}
\alias{pubweb_recode_hceres}
\title{Recode HCERES field into meaningful labels}
\usage{
pubweb_recode_hceres(y)
}
\arguments{
\item{y}{hceres field values in numerical encoding}
}
\value{
factor vector
}
\description{
Recode HCERES field into meaningful labels
}
