% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\name{pubweb_set_config}
\alias{pubweb_set_config}
\title{Define config values}
\usage{
pubweb_set_config(...)
}
\arguments{
\item{...}{named value}
}
\description{
Define config values
}
\examples{
 \dontrun{pubweb_set_config(root_url="https://pubweb.fr")}
}
