% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{pubweb_teams}
\alias{pubweb_teams}
\title{Extract teams attached of each publication of the provided db}
\usage{
pubweb_teams(.db, recode = TRUE)
}
\arguments{
\item{.db}{data.frame() of publication (returned by pubweb_load_db() or a subset)}

\item{recode}{if TRUE numerical team id will be recoded to factor}
}
\value{
data.frame(pub_id, team)
}
\description{
Extract teams attached of each publication of the provided db
}
