% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{recode_author_name}
\alias{recode_author_name}
\title{Recode an author name to full name
This function works with data.frame}
\usage{
recode_author_name(.row)
}
\arguments{
\item{.row}{author data raw}
}
\description{
Recode an author name to full name
This function works with data.frame
}
