% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{recode_authors}
\alias{recode_authors}
\title{Compute authors full name from authors data.frame}
\usage{
recode_authors(aa)
}
\arguments{
\item{aa}{authors data.frame()}
}
\description{
Compute authors full name from authors data.frame
}
