% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{recode_team}
\alias{recode_team}
\title{Recode team to factor (team acronym) from numerical value}
\usage{
recode_team(x)
}
\arguments{
\item{x}{team value as numerical code}
}
\value{
factor with team label as level
}
\description{
Recode team to factor (team acronym) from numerical value
}
