% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team_combinations}
\alias{team_combinations}
\title{Compute Matrix of team combinations in the provided db}
\usage{
team_combinations(.db, without_team = NULL)
}
\arguments{
\item{.db}{pubweb database}

\item{without_team}{numercical code of team to exclude from the matrix}
}
\description{
Compute Matrix of team combinations in the provided db
}
