% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\docType{data}
\name{config}
\alias{config}
\title{Configuration
\describe{
\item{root_url}{Root of the pubweb database site}
\item{teams}{vector of int, List of teams with numerical code as value, and team label as name}
}}
\format{
An object of class \code{environment} of length 3.
}
\usage{
config
}
\description{
Configuration
\describe{
\item{root_url}{Root of the pubweb database site}
\item{teams}{vector of int, List of teams with numerical code as value, and team label as name}
}
}
\keyword{datasets}
