% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\docType{package}
\name{iplespBib-package}
\alias{iplespBib}
\alias{iplespBib-package}
\title{IplespBib package
Load Bibliography database from Iplesp Database}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Turbelin Clément \email{clement.turbelin@upmc.fr} (\href{https://orcid.org/0000-0002-5074-9541}{ORCID})

}
