% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{pubweb_load_db}
\alias{pubweb_load_db}
\title{Load Pubweb databases}
\usage{
pubweb_load_db(ttl = 3600L * 3L)
}
\arguments{
\item{ttl}{int cache time (if 0 will disable cache)}
}
\description{
Load Pubweb databases
}
